requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.27
langchain-text-splitters==0.0.1
langchain-openai==0.0.8
pypdf>=4.0.1
langchain-pinecone>=0.0.1
pinecone-client==3.0.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9